==> Wright/AllRegions/RMarkdowns/02_exploratory_analysis.Rmd <==
---
title: 'Exploratory Analysis'
output:
  html_document:
    code_folding: 'hide'
---

```{r load_packages, warning=FALSE, message=FALSE}
library(tidyverse) ; library(reshape2) ; library(glue) ; library(plotly) ; library(plotlyutils)
library(RColorBrewer) ; library(viridis) ; require(gridExtra) ; library(GGally) ; library(ggpubr)
library(Rtsne)
library(expss)
library(ClusterR)
library(DESeq2) ; library(biomaRt)
library(knitr)
```

Load preprocessed dataset (preprocessing code in 01_data_preprocessing.Rmd)

```{r load_dataset, echo=TRUE, message=FALSE, warning=FALSE}

# Gandal dataset
load('./../Data/preprocessed_data.RData')
datExpr = datExpr %>% data.frame
rownames(datExpr) = datGenes$ensembl_gene_id
DE_info = DE_info %>% data.frame
datMeta = datMeta %>% mutate(ID = title)

# GO Neuronal annotations: regex 'neuron' in GO functional annotations and label the genes that make a match as neuronal
GO_annotations = read.csv('./../Data/genes_GO_annotations.csv')
GO_neuronal = GO_annotations %>% filter(grepl('neuron', go_term)) %>% 
              mutate('ID'=as.character(ensembl_gene_id)) %>% 
              dplyr::select(-ensembl_gene_id) %>% distinct(ID) %>%
              mutate('Neuronal'=1)

# Update DE_info with Neuronal information
DE_info = DE_info %>% mutate('entrezgene'=rownames(.) %>% as.numeric) %>% 
          dplyr::rename('padj' = adj.P.Val, 'log2FoldChange' = logFC) %>%
          left_join(datGenes %>% dplyr::select(entrezgene, ensembl_gene_id) %>% 
                    dplyr::rename('ID' = ensembl_gene_id), by = 'entrezgene') %>%
          left_join(GO_neuronal, by='ID') %>%
          mutate(Neuronal=ifelse(is.na(Neuronal), 0, Neuronal)) %>%
          mutate(significant=padj<0.05 & !is.na(padj))


rm(GO_annotations)
```
<br>

## Mean Level of Expression
<br>

### All samples together
<br>

- There seems to be a small group of genes with lower mean expression than most genes

- There are some samples with lower level of expression than most samples

```{r, fig.width=10}

plot_data = data.frame('ID'=rownames(datExpr), 'Mean'=rowMeans(datExpr))

p1 = plot_data %>% ggplot(aes(Mean)) + geom_density(color='#0099cc', fill='#0099cc', alpha=0.3) +
     xlab('Mean Expression') + ylab('Density') + ggtitle('Mean Expression distribution by Gene') +
     theme_minimal()

plot_data = data.frame('ID'=colnames(datExpr), 'Mean'=colMeans(datExpr))

p2 = plot_data %>% ggplot(aes(Mean)) + geom_density(color='#0099cc', fill='#0099cc', alpha=0.3) +
     xlab('Mean Expression') + ylab('Density') +
     theme_minimal() + ggtitle('Mean expression distribution by Sample')

grid.arrange(p1, p2, nrow=1)


rm(p1, p2, plot_data)
```

### Grouping samples by Phenotype
<br>

The differences in level of expression between Phenotype information are not statistically significant

```{r, fig.width=6, fig.height=4, warning=FALSE, message=FALSE}

plot_data = data.frame('ID'=colnames(datExpr), 'Mean'=colMeans(datExpr)) %>% left_join(datMeta, by='ID')

p1 = plot_data %>% ggplot(aes(Ethnicity, Mean, fill = Ethnicity)) + 
     geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3) + 
     stat_compare_means(label = 'p.signif', method = 't.test', method.args = list(var.equal = FALSE)) +
     xlab('Batch') + ylab('Mean Expression') + theme_minimal() + theme(legend.position = 'none')

p2 = plot_data %>% ggplot(aes(Sex, Mean, fill = Sex)) + 
     geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3) + 
     stat_compare_means(label = 'p.signif', method = 't.test', method.args = list(var.equal = FALSE)) +
     xlab('Gender') + ylab('') + theme_minimal() + theme(legend.position = 'none')

grid.arrange(p1,p2, nrow = 1)


rm(p1,p2)
```

### Grouping genes by Neuronal tag and samples by Diagnosis
<br>

- The two groups of genes seem to be partially characterised by genes with Neuronal function

- In general, the ASD group has a lower mean than the control group (opposite to Gandal and Gupta's results)

- Only the differences in mean expression between Neuronal and non-neuronal genes are statistically significant

```{r GO_annotations, warning=FALSE, fig.width=10, fig.height=8}

plot_data = data.frame('ID'=rownames(datExpr), 'Mean'=rowMeans(datExpr)) %>% 
            left_join(GO_neuronal, by='ID') %>% mutate('Neuronal'=ifelse(is.na(Neuronal),F,T))

p1 = plot_data %>% ggplot(aes(Mean, color=Neuronal, fill=Neuronal)) + geom_density(alpha=0.3) +
                   theme_minimal() + theme(legend.position='bottom') + 
                   ggtitle('Mean expression by gene')

p3 = plot_data %>% ggplot(aes(Neuronal, Mean, fill = Neuronal)) + 
                   geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3) + 
                   stat_compare_means(label = 'p.signif', method = 't.test', 
                                      method.args = list(var.equal = FALSE)) + theme_minimal() + 
                   ylab('Mean Expression') + theme(legend.position = 'none')

plot_data = data.frame('ID'=colnames(datExpr), 'Mean'=colMeans(datExpr)) %>% left_join(datMeta, by='ID')

p2 = plot_data %>% ggplot(aes(Mean, color=Diagnosis, fill=Diagnosis)) + geom_density(alpha=0.3) +
                   theme_minimal() + theme(legend.position='bottom') + 
                   ggtitle('Mean expression by Sample')

p4 = plot_data %>% ggplot(aes(Diagnosis, Mean, fill = Diagnosis)) + 
                   geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3) + 
                   stat_compare_means(label = 'p.signif', method = 't.test', 
                                      method.args = list(var.equal = FALSE)) + theme_minimal() +
                   ylab('Mean Expression') + theme(legend.position = 'none')


grid.arrange(p1, p2, p3, p4, nrow=2)

rm(GO_annotations, plot_data, p1, p2, p3, p4)
```
<br>

### Grouping genes and samples by Diagnosis

In general there doesn't seem to be a lot of variance in mean expression between autism and control samples by gene

```{r asd_vs_ctl, warning=FALSE}
plot_data = data.frame('ID'=rownames(datExpr),
                       'ASD'=rowMeans(datExpr[,datMeta$Diagnosis=='ASD']),
                       'CTL'=rowMeans(datExpr[,datMeta$Diagnosis!='ASD'])) %>%
                       mutate(diff=ASD-CTL, abs_diff = abs(ASD-CTL)) %>%
                       mutate(std_diff = (diff-mean(diff))/sd(diff), distance = abs((diff-mean(diff))/sd(diff)))

plot_data %>% ggplot(aes(ASD, CTL, color = distance)) + geom_point(alpha = plot_data$abs_diff) + 
              geom_abline(color = 'gray') + scale_color_viridis(direction = -1) + 
              ggtitle('Mean expression ASD vs CTL') + theme_minimal() + coord_fixed()
```

```{r}
summary(plot_data$std_diff)
#cat(paste0('Outlier genes: ', paste(plot_data$ID[abs(plot_data$std_diff)>3], collapse = ', ')))
```
There are `r sum(plot_data$distance>3)` genes with a difference between Diagnoses larger than 3 SD to the distance distribution of all genes. Gene `r plot_data$ID[which.max(plot_data$distance)] %>% as.character` has the largest difference in mean expression between ASD and CTL

<br>

- There doesn't seem to be a noticeable difference between mean expression by gene between Diagnosis groups

- Samples with autism tend to have lower values than the control group (as we had already seen above)

```{r, fig.width=10}
plot_data = rbind(data.frame('Mean'=rowMeans(datExpr[,datMeta$Diagnosis=='ASD']), 'Diagnosis'='ASD'),
                  data.frame('Mean'=rowMeans(datExpr[,datMeta$Diagnosis!='ASD']), 'Diagnosis'='CTL')) %>%
            mutate('Diagnosis'=factor(Diagnosis, levels=c('CTL','ASD')))
p1 = ggplotly(plot_data %>% ggplot(aes(Mean, color=Diagnosis, fill=Diagnosis)) + 
              geom_density(alpha=0.3) + theme_minimal())

plot_data = rbind(data.frame('Mean'=colMeans(datExpr[,datMeta$Diagnosis=='ASD']), 'Diagnosis'='ASD'),
                  data.frame('Mean'=colMeans(datExpr[,datMeta$Diagnosis!='ASD']), 'Diagnosis'='CTL')) %>%
            mutate('Diagnosis'=factor(Diagnosis, levels=c('CTL','ASD')))
p2 = ggplotly(plot_data %>% ggplot(aes(Mean, color=Diagnosis, fill=Diagnosis)) + 
              geom_density(alpha=0.3) + theme_minimal() +
              ggtitle('Mean expression by Gene (left) and by Sample (right) grouped by Diagnosis'))

subplot(p1, p2, nrows=1)

rm(p1, p2, plot_data)
```
<br><br>

---

## Visualisations
<br>

### Samples
<br>

#### PCA
<br>

Samples seems to separate relatively well by Diagnosis

ASD samples seem to be more evenly spread out than the Control samples
```{r pca_samples, warning=FALSE, message=FALSE}

pca = datExpr %>% t %>% prcomp

plot_data = data.frame('ID'=colnames(datExpr), 'PC1' = pca$x[,1], 'PC2' = pca$x[,2]) %>% 
            left_join(datMeta, by='ID') %>% dplyr::select('ID','PC1','PC2','Diagnosis') %>% 
            mutate('Diagnosis'=factor(Diagnosis, levels=c('CTL','ASD')))

plot_data %>% ggplot(aes(PC1, PC2, color=Diagnosis)) + geom_point(alpha = 0.8) +
              xlab(paste0('PC1 (',round(100*summary(pca)$importance[2,1],1),'%)')) +
              ylab(paste0('PC2 (',round(100*summary(pca)$importance[2,2],1),'%)')) +
              theme_minimal() + ggtitle('PCA of Samples')


rm(pca, plot_data)
```
<br>

#### MDS
<br>

Looks exactly the same as the PCA visualisation, just inverting the both axes
```{r mds_samples, message=FALSE, warning=FALSE}

d = datExpr %>% t %>% dist
fit = cmdscale(d, k=2)

plot_data = data.frame('ID'=colnames(datExpr), 'C1'=fit[,1], 'C2'=fit[,2]) %>% left_join(datMeta, by='ID') %>% 
            dplyr::select('C1','C2','Diagnosis') %>%
            mutate('Diagnosis'=factor(Diagnosis, levels=c('CTL','ASD')))

plot_data %>% ggplot(aes(C1, C2, color=Diagnosis)) + geom_point(alpha = 0.8) + theme_minimal() + ggtitle('MDS')

rm(d, fit, plot_data)
```
<br>

#### t-SNE
<br>

T-SNE seems to be struggling to separate the samples by Diagnosis

Using a perplexity of 2 the ASD samples seem to gather in the center and Controls in the periphery
```{r tsne_samples, warning=FALSE, message=FALSE, fig.width=10, fig.height=10}

perplexities = c(1,2,5,10)
ps = list()

for(i in 1:length(perplexities)){
  set.seed(123)
  tsne = datExpr %>% t %>% Rtsne(perplexity=perplexities[i])
  plot_data = data.frame('ID'=colnames(datExpr), 'C1'=tsne$Y[,1], 'C2'=tsne$Y[,2]) %>% 
              left_join(datMeta, by='ID') %>%
              dplyr::select('C1','C2','Diagnosis') %>%
              mutate('Diagnosis'=factor(Diagnosis, levels=c('CTL','ASD')))
  ps[[i]] = plot_data %>% ggplot(aes(C1, C2, color=Diagnosis)) + geom_point() + theme_minimal() +
            ggtitle(paste0('Perplexity=',perplexities[i])) + theme(legend.position='none')
}

grid.arrange(grobs=ps, nrow=2)


rm(ps, perplexities, tsne, i)
```
<br>

### Genes
<br>

#### PCA
<br>

- The First Principal Component explains over 97% of the total variance

- There's a really strong correlation between the mean expression of a gene and the 1st principal component

- The magnitude of the second principal component seems to be related to the level of expression of the genes (this didn't happen in the other datasets)

```{r pca_genes, fig.width=10, warning=FALSE, message=FALSE}
pca = datExpr %>% prcomp

plot_data = data.frame( 'PC1' = pca$x[,1], 'PC2' = pca$x[,2], 'MeanExpr'=rowMeans(datExpr))

plot_data %>% ggplot(aes(PC1, PC2, color=MeanExpr)) + geom_point(alpha=0.3) + theme_minimal() + 
              scale_color_viridis() + ggtitle('PCA') +
              xlab(paste0('PC1 (',round(100*summary(pca)$importance[2,1],1),'%)')) +
              ylab(paste0('PC2 (',round(100*summary(pca)$importance[2,2],1),'%)'))

rm(pca, plot_data)
```
<br>

#### t-SNE
<br>

Higher perplexities capture a cleaner visualisation of the data ordered by mean expression, in a similar (although not as linear) way to PCA
```{r tsne_genes, fig.width=10, fig.height=7}

perplexities = c(1,2,5,10,50,100)
ps = list()

for(i in 1:length(perplexities)){
  tsne = read.csv(paste0('./../Visualisations/tsne_perplexity_',perplexities[i],'.csv'))
  plot_data = data.frame('C1'=tsne[,1], 'C2'=tsne[,2], 'MeanExpr'=rowMeans(datExpr))
  ps[[i]] = plot_data %>% ggplot(aes(C1, C2, color=MeanExpr)) + geom_point(alpha=0.3) + theme_minimal() +
            scale_color_viridis() + ggtitle(paste0('Perplexity = ',perplexities[i])) + theme(legend.position='none')
}

grid.arrange(grobs=ps, nrow=2)

rm(perplexities, ps, tsne, i)
```

<br>

---

<br>

## Differential Expression Analysis
<br>

**Only `r sum(DE_info$padj<0.05, na.rm=T)` genes (~`r round(mean(DE_info$padj<0.05, na.rm=T)*100,4)`% vs ~28% in Gandal’s dataset) are significant using a threshold of 0.05 for the adjusted p-value and a without a log Fold Change threshold (keeping the null hypothesis $H_0: LFC=0$)**

```{r DE_count}
table(DE_info$padj<0.05, useNA='ifany')
```

```{r logFC_vs_adjPval, fig.width=10, warning=FALSE, message=FALSE}

p = DE_info %>% ggplot(aes(log2FoldChange, padj, color=significant)) + geom_point(alpha=0.2) + 
    scale_y_sqrt() + xlab('log2 Fold Change') + ylab('Adjusted p-value') + theme_minimal()
ggExtra::ggMarginal(p, type = 'density', color='gray', fill='gray', size=10)


rm(p)
```

- There is a negative relation between LFC and mean expression

- The relation is strongest for genes with low levels of expression

```{r logFC_vs_meanExpr, fig.width=10, warning=FALSE, message=FALSE}

plot_data = data.frame('ID'=rownames(datExpr), 'meanExpr'=rowMeans(datExpr)) %>% left_join(DE_info, by='ID')

plot_data %>% ggplot(aes(meanExpr, abs(log2FoldChange))) + 
              geom_point(alpha = 0.3, color='#0099cc') + geom_smooth(method='lm', color = 'gray') + 
              theme_minimal() + scale_y_sqrt() + theme(legend.position = 'bottom') +
              xlab('Mean Expression') + ylab('LFC Magnitude') + 
              ggtitle('Log fold change by level of expression')

```

List of DE genes

```{r}
# Get genes names
getinfo = c('ensembl_gene_id','external_gene_id')
mart = useMart(biomart='ENSEMBL_MART_ENSEMBL', dataset='hsapiens_gene_ensembl', host='feb2014.archive.ensembl.org')
gene_names = getBM(attributes=getinfo, filters=c('ensembl_gene_id'), values=plot_data$ID, mart=mart) %>% 
             rename(external_gene_id = 'gene_name', ensembl_gene_id = 'ID')

top_genes = plot_data %>% left_join(gene_names, by='ID') %>% filter(padj<0.05) %>% arrange(-abs(log2FoldChange))

kable(top_genes %>% dplyr::select(ID, gene_name, log2FoldChange, padj, Neuronal))


rm(top_genes)
```
<br><br>

## Effects of modifying the log fold change threshold
<br>

No point in doing this having only 2 DE genes

<br><br>

---

#### Session info

```{r}
sessionInfo()
```
<br><br>
