==> Wright/AllRegions/RMarkdowns/06_modules.Rmd <==
---
title: 'WGCNA Exploratory Analysis'
output:
  html_document:
    code_folding: 'hide'
---

```{r load_packages, warning=FALSE, message=FALSE}

library(tidyverse) ; library(reshape2) ; library(glue) ; library(plotly) ; library(dendextend)
library(RColorBrewer) ; library(viridis) ; require(gridExtra) ; library(GGally) ; library(ggpubr) ; library(ggExtra)
library(WGCNA)
library(expss)
library(polycor)
library(biomaRt)
library(clusterProfiler) # ORA
library(foreach) ; library(doParallel)

SFARI_colour_hue = function(r) {
  pal = c('#FF7631','#FFB100','#E8E328','#8CC83F','#62CCA6','#59B9C9','#b3b3b3','#808080','gray','#d9d9d9')[r]
}
```

Load preprocessed dataset (preprocessing code in01_data_preprocessing.Rmd) and clustering (pipeline in 05_WGCNA.Rmd)
```{r load_dataset, message=FALSE, warning=FALSE}

# Gandal dataset
load('./../Data/preprocessed_data.RData')
datExpr = datExpr %>% data.frame
rownames(datExpr) = datGenes$ensembl_gene_id
DE_info = DE_info %>% data.frame
datMeta = datMeta %>% mutate(ID = title)


# GO Neuronal annotations: regex 'neuron' in GO functional annotations and label the genes that make a match as neuronal
GO_annotations = read.csv('./../Data/genes_GO_annotations.csv')
GO_neuronal = GO_annotations %>% filter(grepl('neuron', go_term)) %>% 
              mutate('ID'=as.character(ensembl_gene_id)) %>% 
              dplyr::select(-ensembl_gene_id) %>% distinct(ID) %>%
              mutate('Neuronal'=1)


# SFARI Genes
SFARI_genes = read_csv('./../../../SFARI/Data/SFARI_genes_01-03-2020_w_ensembl_IDs.csv')
SFARI_genes = SFARI_genes[!duplicated(SFARI_genes$ID) & !is.na(SFARI_genes$ID),]


# Clusterings
clusterings = read_csv('./../Data/clusters.csv')


# Update DE_info with SFARI and Neuronal information
genes_info = DE_info %>% mutate('entrezgene'=rownames(.) %>% as.numeric) %>% 
             dplyr::rename('padj' = adj.P.Val, 'log2FoldChange' = logFC) %>%
             left_join(datGenes %>% dplyr::select(entrezgene, ensembl_gene_id) %>% 
                       dplyr::rename('ID' = ensembl_gene_id), by = 'entrezgene') %>% 
             left_join(SFARI_genes, by='ID') %>% 
             mutate(`gene-score`=ifelse(is.na(`gene-score`), 'Others', `gene-score`)) %>%
             distinct(ID, .keep_all = TRUE) %>% left_join(GO_neuronal, by='ID') %>%
             mutate(Neuronal=ifelse(is.na(Neuronal), 0, Neuronal)) %>%
             mutate(gene.score=ifelse(`gene-score`=='Others' & Neuronal==1, 'Neuronal', `gene-score`), 
                    significant=padj<0.05 & !is.na(padj)) %>%
             mutate(Group = factor(ifelse(gene.score %in% c('Neuronal','Others'), gene.score, 'SFARI'), 
                    levels = c('SFARI', 'Neuronal', 'Others'))) %>% left_join(clusterings, by='ID')


rm(DE_info, GO_annotations, clusterings, efit)
```

---

# Dynamic Tree vs Dyamic Hybrid
<br>

### Comparing module sizes
<br>

Dynamic Tree creates less and more balanced modules (opposite to Gandal's modules)
```{r, fig.width=10}

plot_data = table(genes_info$DynamicTree) %>% data.frame %>% arrange(desc(Freq)) %>% filter(Var1 != 'gray')

p1 = plot_data %>% ggplot(aes(x=reorder(Var1, -Freq), y=Freq)) + 
     geom_bar(stat='identity', fill=plot_data$Var1) + 
     geom_hline(yintercept = mean(plot_data$Freq), color = 'gray',linetype = 'dashed') +
     ggtitle('Modules using Dynamic Tree') + ylab('Number of genes') + xlab('Module') + 
     theme_minimal() + theme(axis.text.x = element_blank())

plot_data = table(genes_info$DynamicHybrid) %>% data.frame %>% arrange(desc(Freq)) %>% filter(Var1 != 'gray')

p2 = plot_data %>% ggplot(aes(x=reorder(Var1, -Freq), y=Freq)) + 
     geom_bar(stat='identity', fill=plot_data$Var1) + 
     geom_hline(yintercept = mean(plot_data$Freq), color = 'gray',linetype = 'dashed') +
     ggtitle('Modules using Dynamic Hybrid') + ylab('Number of genes') + xlab('Module') + 
     theme_minimal() + theme(axis.text.x = element_blank())

grid.arrange(p1,p2, nrow=2, top = 'Comparison of Module sizes')

rm(p1, p2, plot_data)
```
<br>

### Comparing genes left without a module
<br>

Dynamic Tree leaves `r sum(genes_info$DynamicTree=='gray')` genes without cluster (`r round(mean(genes_info$DynamicTree=='gray')*100)`%)
Dynamic Hybrid leaves `r sum(genes_info$DynamicHybrid=='gray')` genes without cluster (`r round(mean(genes_info$DynamicHybrid=='gray')*100)`%)
<br>

Dynamic Tree leaves many more genes without a cluster (too many). As we saw in the dendrograms in 05_WGCNA.html, the genes usually left without a cluster are the ones closest to the root of the dendrogram, now we want to see if they share some characteristic in common

- The group of genes with low level of expression and large values in the 2nd principal component seem to be left without a cluster more often than the rest of the genes. We had seen that this genes were noisy in the preprocessing pipeline, so it's not so bad that we are losing them

- Genes assigned to a cluster also seem to have a higher level of expression than the genes left unassigned (becaues they have higher PC1 values)

```{r pca_plots, warning=FALSE, fig.align='center'}

pca = datExpr %>% prcomp

plot_data = data.frame('ID'=rownames(datExpr), 'PC1' = pca$x[,1], 'PC2' = pca$x[,2]) %>%
            left_join(genes_info, by='ID') %>% mutate('hasCluster' = DynamicTree!='gray', 
                                                      'hasSFARIScore' = `gene-score`!='Others') %>%
            apply_labels(`gene-score` = 'SFARI Gene score', DynamicTree = 'Dynamic Tree Algorithm', 
                         significant = 'Differentially Expressed', hasCluster = 'Belongs to a Module',
                         hasSFARIScore = 'Has a SFARI Score', syndromic = 'Has syndromic tag') %>%
            mutate(alpha_cluster = ifelse(hasCluster==TRUE, 0.8, 0.1),
                   alpha_signif = ifelse(significant==TRUE, 0.8, 0.1))

p = plot_data %>% ggplot(aes(PC1, PC2, color=hasCluster)) + geom_point(alpha = plot_data$alpha_cluster) + 
    theme_minimal() + ggtitle('Genes are assigned to a cluster') + theme(legend.position='bottom')

ggExtra::ggMarginal(p, type = 'density', groupColour = TRUE, size = 10)


rm(p)
```
<br>

**Most of the genes with a SFARI score are not assigned to a cluster:**
<br>

`r sum(plot_data$hasSFARIScore & !plot_data$hasCluster)` of the SFARI genes (~`r round(100*sum(plot_data$hasSFARIScore & !plot_data$hasCluster)/sum(plot_data$hasSFARIScore))`%) are not assigned to any cluster

```{r}
cro(plot_data$hasSFARIScore, list(plot_data$hasCluster, total()))
```
<br>

## Conclusion:
<br>

The main difference between algorithms is that Dynamic Hybrid clusters outlier genes and Dynamic Tree leaves them out, so Dynamic Tree would give me a 'cleaner' group of genes to work with, but at the expense of losing too many genes

I think in this case, using the Dynamic Tree modules wouldn't be feasible given the huge ammount of unassigned genes, so I'm going to use the Dynamic Hybrid algorithm to keep more genes

```{r select_clustering_method}
clustering_selected = 'DynamicHybrid'
genes_info$Module = genes_info[,clustering_selected]
```

---
---

<br>

# Dynamic Hybrid Modules

<br>

*The colour of the modules is the arbitrary one assigned during the WGCNA algorithm, where the gray cluster actually represents all the genes that were left without a cluster (so it's not actually a cluster).
<br>

The Dynamic Hybrid algorithm created `r length(unique(genes_info$Module))-1` modules and leaves `r sum(genes_info$Module=='gray')` genes without a module (this is still quite large)
```{r dynamic_hybrid_module_count, fig.width=10}

table(genes_info$Module)

plot_data = table(genes_info$Module) %>% data.frame %>% arrange(desc(Freq))

ggplotly(plot_data %>% ggplot(aes(x=reorder(Var1, -Freq), y=Freq)) + geom_bar(stat='identity', fill=plot_data$Var1) + 
         ggtitle('Module size') + ylab('Number of genes') + xlab('Module') + theme_minimal() + 
         theme(axis.text.x = element_text(angle = 90)))
```

<br>

## Relation to external clinical traits
<br>

### Quantifying module-trait associations
<br>

In the WGCNA documentation they use Pearson correlation to calculate correlations, I think all of their variables were continuous. Since I have categorical variables I'm going to use the `hetcor` function, that calculates Pearson, polyserial or polychoric correlations depending on the type of variables involved.

- I'm not sure how the `corPvalueStudent` function calculates the p-values and I cannot find any documentation...

- Compared correlations using Pearson correlation and with hetcor and they are very similar, but a bit more extreme with hetcor. The same thing happens with the p-values.

```{r calc_module_trait_associations, message=FALSE, warning=FALSE}

datTraits = datMeta %>% dplyr::select(Diagnosis, Sex, Age)

# Recalculate MEs with color labels
ME_object = datExpr %>% t %>% moduleEigengenes(colors = genes_info$Module)
MEs = orderMEs(ME_object$eigengenes)

# Calculate correlation between eigengenes and the traits and their p-values
moduleTraitCor = MEs %>% apply(2, function(x) hetcor(x, datTraits)$correlations[1,-1]) %>% t
rownames(moduleTraitCor) = colnames(MEs)
colnames(moduleTraitCor) = colnames(datTraits)
moduleTraitPvalue = corPvalueStudent(moduleTraitCor, nrow(datExpr))

# Create text matrix for the Heatmap
textMatrix = paste0(signif(moduleTraitCor, 2), ' (', signif(moduleTraitPvalue, 1), ')')
dim(textMatrix) = dim(moduleTraitCor)

# In case there are any NAs
if(sum(!complete.cases(moduleTraitCor))>0){
  print(paste0(sum(is.na(moduleTraitCor)),' correlation(s) could not be calculated')) 
}


rm(ME_object)
```

- In general, modules have strong correlations with Diagnosis with really small p-values and not much relation with anything else

```{r plot_heatmap, fig.width=10, fig.height=25, warning=FALSE, message=FALSE}

# Sort moduleTraitCor by Diagnosis
moduleTraitCor = moduleTraitCor[order(moduleTraitCor[,1], decreasing=TRUE),]
moduleTraitPvalue = moduleTraitPvalue[order(moduleTraitCor[,1], decreasing=TRUE),]

# Create text matrix for the Heatmap
textMatrix = paste0(signif(moduleTraitCor, 2), ' (', signif(moduleTraitPvalue, 1), ')')
dim(textMatrix) = dim(moduleTraitCor)


labeledHeatmap(Matrix = moduleTraitCor, xLabels = names(datTraits), 
               yLabels =  gsub('ME','',rownames(moduleTraitCor)), 
               yColorWidth=0, colors = brewer.pal(11,'PiYG'), 
               bg.lab.y = gsub('ME','',rownames(moduleTraitCor)),
               textMatrix = textMatrix, setStdMargins = FALSE, cex.text = 0.8, 
               cex.lab.y = 0.75, zlim = c(-1,1), main = paste('Module-Trait relationships'))

diagnosis_cor = data.frame('Module' = gsub('ME','',rownames(moduleTraitCor)),
                           'MTcor' = moduleTraitCor[,'Diagnosis'],
                           'MTpval' = moduleTraitPvalue[,'Diagnosis'])

genes_info = genes_info %>% left_join(diagnosis_cor, by='Module')


rm(moduleTraitCor, moduleTraitPvalue, textMatrix, diagnosis_cor)
```

**Modules with a high Module-Diagnosis (absolute) correlation should have a high content of differentially expressed genes:**

Not enough DE genes to test this

<br>

## Gene Significance and Module Membership
<br>

**Gene significance:** is the value between the correlation between the gene and the trait we are interested in. A positive gene significance means the gene is overexpressed and a negative value means its underexpressed. (The term 'significance' is not very acurate because it's not actually measuring statistical significance, it's just a correlation, but that's how they call it in WGCNA...)

**Module Membership** is the correlation of the module's eigengene and the expression profile of a gene. The higher the Module Membership, the more similar the gene is to the genes that constitute the module. (I won't use this measure yet)

<br>

**Note:** Some correlations weren't able to be calculated with the original function (weirdly because the correlation was too strong), so they are calculated again in the end using the `polyserial` function instead of `hetcor`, which is robust enough to calculate them
```{r module_membership}
# It's more efficient to iterate the correlations one by one, otherwise it calculates correlations between the eigengenes and also between the genes, which we don't need

# Check if MM information already exists and if not, calculate it
if(file.exists(paste0('./../Data/dataset_', clustering_selected, '.csv'))){
  
  dataset = read.csv(paste0('./../Data/dataset_', clustering_selected, '.csv'))
  dataset$Module = dataset[,clustering_selected]
  
} else {
  
  ############# 1. Calculate Gene Significance
  GS_info = data.frame('ID'=rownames(datExpr),
                       'GS'=datExpr %>% apply(1, function(x) hetcor(x,datMeta$Diagnosis)$correlations[1,2])) %>%
            mutate('GSpval'=corPvalueStudent(GS, ncol(datExpr)))
  
  #############  2. Calculate Module Membership
  
  #setup parallel backend to use many processors
  cores = detectCores()
  cl = makeCluster(cores-1)
  registerDoParallel(cl)
  
  # Create matrix with MM by gene
  MM = foreach(i=1:nrow(datExpr), .combine=rbind) %dopar% {
    library(polycor)
    tempMatrix = apply(MEs, 2, function(x) hetcor(as.numeric(datExpr[i,]), x)$correlations[1,2])
    tempMatrix
  }
  
  # Stop clusters
  stopCluster(cl)
  
  rownames(MM) = rownames(datExpr)
  colnames(MM) = paste0('MM',gsub('ME','',colnames(MEs)))
  
  # Calculate p-values
  MMpval = MM %>% corPvalueStudent(ncol(datExpr)) %>% as.data.frame
  colnames(MMpval) = paste0('MMpval', gsub('ME','',colnames(MEs)))
  
  MM = MM %>% as.data.frame %>% mutate(ID = rownames(.))
  MMpval = MMpval %>% as.data.frame %>% mutate(ID = rownames(.))
  
  # Join and save results
  dataset = genes_info %>% dplyr::select(ID, `gene-score`, clustering_selected, MTcor, MTpval) %>%
            left_join(GS_info, by='ID') %>%
            left_join(MM, by='ID') %>%
            left_join(MMpval, by='ID')
  
  write.csv(dataset, file = paste0('./../Data/dataset_', clustering_selected, '.csv'), row.names = FALSE)
  
  rm(cores, cl) 
  
}

GS_missing = dataset$ID[is.na(dataset$GS)] %>% as.character

if(length(GS_missing)>0){
  
  cat(paste0(length(GS_missing),' correlations between genes and Diagnosis could not be calculated, ',
             'calculating them with the polyserial function'))
  
  for(g in GS_missing){
    dataset$GS[dataset$ID == g] = polyserial(as.numeric(datExpr[g,]), datMeta$Diagnosis)
  }
  
}

rm(GS_missing)

```
<br>

## Analysing concordance between these metrics in the genes

<br>

### 1. Gene Significance vs Log Fold Change
<br>

Gene significance and Log Fold Chance are two different ways to measure the same thing, so there should be a concordance between them

Both variables agree with each other relatively well

```{r GS_lfc, warning=FALSE, message=FALSE}

plot_data = dataset %>% dplyr::select(ID, MTcor, GS) %>% 
            left_join(genes_info %>% dplyr::select(ID, gene.score), by='ID') %>%
            left_join(genes_info %>% dplyr::select(ID, AveExpr, log2FoldChange, significant, Module), 
                      by='ID') %>%
            left_join(data.frame(MTcor=unique(dataset$MTcor)) %>% arrange(by=MTcor) %>% 
                                 mutate(order=1:length(unique(dataset$MTcor))), by='MTcor')

ggplotly(plot_data %>% ggplot(aes(GS, log2FoldChange)) +
         geom_point(color=plot_data$Module, alpha=0.5, aes(ID=Module)) + 
         geom_smooth(color='gray', se=FALSE) + theme_minimal() + xlab('Gene Significance') + 
         ggtitle(paste0('R^2 = ', round(cor(plot_data$log2FoldChange, plot_data$GS)[1]**2, 2))))

```
<br>

### 2. Module-Diagnosis correlation vs Gene Significance
<br>

In general, modules with the highest Module-Diagnosis correlation should have genes with high Gene Significance

**Note:** For the Module-Diagnosis plots, if you do boxplots, you lose the exact module-diagnosis correlation and you only keep the order, so I decided to compensate this downside with a second plot, where each point is plotted individually using their module's Module-Diagnosis correlation as the x axis. I think the boxplot plot is easier to understand but the second plot contains more information, so I don't know which one is better.

```{r MD_GS_boxplot, fig.width=12}

plot_data = plot_data %>% arrange(order)

ggplotly(plot_data %>% ggplot(aes(order, GS, group=order)) + 
         geom_hline(yintercept=0, color='gray', linetype='dotted') +
         geom_boxplot(fill=unique(plot_data$Module)) + theme_minimal() + 
         xlab('Modules ordered by Module-Diagnosis correlation') + ylab('Gene Significance'))

```

```{r MD_GS_points, fig.width=10, warning=FALSE, message=FALSE}
plot_data %>% ggplot(aes(MTcor, GS)) + geom_hline(yintercept=0, color='gray', linetype='dotted') + 
         geom_point(color=plot_data$Module, alpha=0.1, aes(id=ID)) + geom_smooth(color='gray', se=FALSE) + 
         theme_minimal() + xlab('Module-Diagnosis correlation') + ylab('Gene Significance') + 
         ggtitle(paste0('R^2=',round(cor(plot_data$MTcor, plot_data$GS)^2,2)))
```
<br>

### 3. Module-Diagnosis correlation vs Log Fold Change
<br>

The same happens with the Log Fold Change

```{r MD_LFC_boxplot, fig.width=12, warning=FALSE, message=FALSE}
ggplotly(plot_data %>% ggplot(aes(order, log2FoldChange, group=order)) + geom_hline(yintercept=0, color='gray', linetype='dotted') +
         geom_boxplot(fill=unique(plot_data$Module)) + 
         theme_minimal() + xlab('Modules ordered by Module-Diagnosis correlation') + ylab('log2FoldChange'))
```

```{r MD_LFC_points, fig.width=10, warning=FALSE, message=FALSE}
ggplotly(plot_data %>% ggplot(aes(MTcor, log2FoldChange)) + 
         geom_hline(yintercept=0, color='gray', linetype='dotted') +
         geom_point(color=plot_data$Module, alpha=0.1, aes(id=ID)) + geom_smooth(color='gray', se=FALSE) + 
         theme_minimal() + xlab('Module-Diagnosis correlation') + ylab('log2FoldChange') + 
         ggtitle(paste0('R^2=',round(cor(plot_data$MTcor, plot_data$log2FoldChange)^2,2))))
```

<br>

### 4. Module-Diagnosis vs Mean Expression
<br>

In theory, there shouldn't be a relation between module-diagnosis and mean expression, but in the the exploratory analysis, we saw that the overexpressed genes tended to have lower levels of expression than the overexpressed genes, and this pattern can be seen in these plots where the modules with negative Module-Diagonsis correlation have slightly higher levels of expression than the modules with positive Module-Diagnosis correlation, although this pattern is note very strong and all modules have similar levels of expression.

```{r MD_ME_boxplot, fig.width=10, warning=FALSE, message=FALSE}
ggplotly(plot_data %>% ggplot(aes(order, AveExpr, group=order)) + 
         geom_hline(yintercept=mean(plot_data$AveExpr), color='gray', linetype='dotted') +
         geom_boxplot(fill=unique(plot_data$Module)) + theme_minimal() + 
         xlab('Modules ordered by Module-Diagnosis correlation') + ylab('Mean Expression'))
```

```{r MD_ME_points, fig.width=10, warning=FALSE, message=FALSE}
plot_data %>% ggplot(aes(MTcor, AveExpr)) + geom_point(alpha=.2, color=plot_data$Module, aes(id=ID)) + 
         geom_hline(yintercept=mean(plot_data$AveExpr), color='gray', linetype='dotted') + 
         geom_smooth(color='gray', alpha=0.3) + theme_minimal() + 
         xlab('Module-Diagnosis correlation') + ylab('Mean Expression') +
         ggtitle(paste0('R^2=',round(cor(plot_data$MTcor, plot_data$AveExpr)^2,2)))
```
<br>

### Conclusion:
<br>

All of the variables seem to agree with each other, Modules with a high correlation with Diagnosis tend to have genes with high values of Log Fold Change as well as high values of Gene Significance, and the gray module, which groups all the genes that weren't assigned to any cluster tends to have a very poor performance in all of the metrics.

<br><br>

---

<br><br>

## SFARI Scores
<br>

Since SFARI scores genes depending on the strength of the evidence linking it to the development of autism, in theory, there should be some concordance between the metrics we have been studying above and these scores...

### SFARI Scores vs Gene Significance
<br>

- The higher the SFARI Gene Score, the lower the Gene Significance, although this pattern is quite subtle

- All SFARI scores have distributions similar to the genes with Neuronal annotations

- These differences are generally not statistically significant

```{r SFARI_GS, warning=FALSE, message=FALSE, fig.height=8, fig.width=6, fig.align='center'}

comparisons = list(c('1','2'), c('2','3'), c('3','Neuronal'), c('Neuronal','Others'),
                   c('1','3'), c('3','Others'), c('2','Neuronal'),
                   c('1','Neuronal'), c('2','Others'), c('1','Others'))
increase = 0.05
base = 0.75
pos_y_comparisons = c(rep(base, 4), rep(base + increase, 2), base + 2:5*increase)

plot_data %>% ggplot(aes(gene.score, abs(GS), fill=gene.score)) + 
              geom_boxplot(outlier.colour='#cccccc', outlier.shape='o', outlier.size=3) + 
              stat_compare_means(comparisons = comparisons, label = 'p.signif', method = 't.test', 
                                 method.args = list(var.equal = FALSE), label.y = pos_y_comparisons, 
                                 tip.length = .01) +
              scale_fill_manual(values=SFARI_colour_hue(r=c(1:3,8,7))) + theme_minimal() + 
              ylab('Gene Significance Magnitude') + xlab('SFARI Scores') + theme(legend.position='none')


rm(increase, base, pos_y_comparisons)
```

### SFARI Scores vs Module-Diagnosis correlation
<br>

- SFARI Score 3 has the highest Module-Diagnosis distribution of all groups, but not by much

- There differences are usually not statistically significant

**Note:** Filtering out gray module because otherwise too many genes share the same Module-Diagnosis correlation
```{r SFARI_MD, warning=FALSE, message=FALSE, fig.height=8, fig.width=6, fig.align='center'}

increase = 0.05
base = 0.95
pos_y_comparisons = c(rep(base, 4), rep(base + increase, 2), base + 2:5*increase)

plot_data %>% filter(Module != 'gray') %>% ggplot(aes(gene.score, abs(MTcor), fill=gene.score)) + 
              geom_boxplot(outlier.colour='#cccccc', outlier.shape='o', outlier.size=3) + 
              stat_compare_means(comparisons = comparisons, label = 'p.signif', method = 't.test', 
                                 method.args = list(var.equal = FALSE), label.y = pos_y_comparisons, 
                                 tip.length = .01) +
              scale_fill_manual(values=SFARI_colour_hue(r=c(1:3,8,7))) + theme_minimal() + 
              ylab('Module-Diagnosis Correlation Magnitude') + xlab('SFARI Scores') + theme(legend.position='none')


rm(increase, base, pos_y_comparisons)
```

### SFARI Enrichment vs Module-Diagnosis correlation
<br>

To study the strength of the relationship between a module and the biological signal found in the data related to ASD we have the Module-Diagnosis metric. We want to do a similar thing to calculate the 'strength' of the presence of SFARI Genes in each module to be able to identify the modules 'enriched' in SFARI Genes
<br>

The easiest metric would be to calculate the percentage of genes in each module that belong to the SFARI Genes, but this metric doesn’t take into account the size of the model and can favour more extreme but less robust results (small modules)
<br>

Instead, we are going to use the Over Representation Analysis (ORA) approach, that uses the hypergeometric distribution to calculate the probability of a module of size $n$ containing at least $k$ SFARI Genes and we are going to interpret this probability as the 'enrichment' of SFARI Genes in each module
<br>

#### Hypergeometric distribution explanation:
<br>
If we interpret the number of genes ($n$) in a cluster as $n$ random draws without replacement from a finite population of size $N$, and the number of SFARI genes in the cluster ($s$) as $s$ successes in those $n$ draws, where we know that $N$ contains exactly $S$ successes, then we can use the hypergeometric distribution to calculate the statistical significance of having drawn $s$ successes out of $n$ draws, and use this value to select the clusters with the highest statistical significance
<br><br>

#### Percentage of SFARI Genes vs ORA by Module
<br>


```{r, warning=FALSE, message=FALSE}
# Calculate % and ORA of SFARI Genes in each module

modules = unique(genes_info$Module[genes_info$Module!='gray']) %>% as.character

# We need the entrez ID of the genes for this
getinfo = c('ensembl_gene_id','entrezgene')
mart = useMart(biomart='ENSEMBL_MART_ENSEMBL', dataset='hsapiens_gene_ensembl', 
               host='feb2014.archive.ensembl.org')
biomart_output = getBM(attributes=getinfo, filters=c('ensembl_gene_id'), 
                       values=genes_info$ID[genes_info$Module!='gray'], mart=mart) %>%
                 left_join(genes_info %>% dplyr::select(ID, Module,`gene-score`), by = c('ensembl_gene_id'='ID'))

# We need to build a term2gene dataframe with the genes and their SFARI Scores
term2gene = biomart_output %>% mutate(term = ifelse(`gene-score` == 'Others', 'Others', 'SFARI'), 
                                      'gene' = entrezgene) %>%  dplyr::select(term, gene) %>% distinct


enrichment_data = data.frame('Module' = modules, 'size' = 0, 'perc_SFARI' = 0,
                             'pval_ORA' = 0, 'padj_ORA' = 0)

for(i in 1:length(modules)){
  module = modules[i]
  genes_in_module = biomart_output$entrezgene[biomart_output$Module==module]
  ORA_module = enricher(gene = genes_in_module, universe = biomart_output$entrezgene %>% as.character, 
                        pAdjustMethod = 'bonferroni', TERM2GENE = term2gene, 
                        pvalueCutoff = 1, qvalueCutoff = 1, maxGSSize = 50000) %>% 
                        data.frame %>% dplyr::select(-geneID,-Description)
  ORA_pval = ifelse('SFARI' %in% ORA_module$ID, ORA_module$pvalue[ORA_module$ID=='SFARI'], 1)
  ORA_padj = ifelse('SFARI' %in% ORA_module$ID, ORA_module$p.adjust[ORA_module$ID=='SFARI'], 1)
  enrichment_data[i,-1] = c(length(genes_in_module), 
                            mean(genes_info$`gene-score`[genes_info$Module==module]!='Others'), 
                            ORA_pval, ORA_padj)
}

enrichment_data = enrichment_data %>% 
                  left_join(genes_info %>% dplyr::select(Module, MTcor) %>% unique, by = 'Module')

rm(i, module, genes_in_module, ORA_module, ORA_pval, ORA_padj, getinfo, mart, term2gene)
```

**Note:** in ORA, a lower probability indicates a higher enrichment in SFARI Genes, so I'm using 1-probability to measure the enrichment of each module
<br>

The two metrics agree quite well, but if we focus on the modules with the highest 'enrichment' with both metrics we can see that the smallest modules are below the trend line while the larger ones are above, which shows the bias towards small modules the percentage of SFARI Genes metric has, making the ORA Enrichment metric more desirable

```{r, fig.align='center', warning=FALSE, message=FALSE}
ggplotly(enrichment_data %>% ggplot(aes(perc_SFARI, 1-pval_ORA, size=size)) + 
         geom_point(color = enrichment_data$Module, alpha = 0.5, aes(id=Module)) + 
         geom_smooth(color='gray', se = FALSE) + 
         xlab('% of SFARI Genes') + ylab('ORA Enrichment') + ggtitle('Modules') + theme_minimal())
```

#### SFARI Enrichment vs Module-Diagnosis Correlation
<br>

There doesn't seem to be a strong relation between these two metrics, perhaps the modules with the largest Module-Diagnosis correlations have lower Enrichment in SFARI Genes than the rest of the dataset

<br>
Cor(SFARI Enrichment, Module-Diagnosis) = `r round(cor(enrichment_data$MTcor, 1-enrichment_data$pval_ORA),2)[1]`
<br>

- Correlation for modules with a positive Module-Diagnosis correlation = `r round(cor(enrichment_data$MTcor[enrichment_data$MTcor>=0], 1-enrichment_data$pval_ORA[enrichment_data$MTcor>=0]),2)[1]`

- Correlation for modules with a negative Module-Diagnosis correlation = `r round(cor(enrichment_data$MTcor[enrichment_data$MTcor<0], 1-enrichment_data$pval_ORA[enrichment_data$MTcor<0]),2)[1]`

```{r, warning=FALSE, message=FALSE, fig.align='center'}
ggplotly(enrichment_data %>% ggplot(aes(MTcor, 1-pval_ORA, size=size)) + 
         geom_point(color=enrichment_data$Module, alpha=0.5, aes(id=Module)) + 
         geom_smooth(color='#cccccc', size = 0.5, alpha=0.1) + xlab('Module-Diagnosis Correlation') + 
         ggtitle('SFARI Enrichment vs Module-Diagnosis Correlation') +
         ylab('Enrichment in SFARI Genes') + theme_minimal() + theme(legend.position = 'none'))

# Save for comparison with old SFARI Genes
enrichment_data_new_SFARI = enrichment_data
```

Dividing the SFARI Genes by score we get the same pattern, just noisier because the groups are smaller now

```{r, warning=FALSE, message=FALSE, fig.align='center'}

# We need to build a term2gene dataframe with the genes and their SFARI Scores
term2gene = biomart_output %>%
            mutate('term' = ifelse(`gene-score` == 'Others', 'Others', `gene-score`), 
                   'gene' = entrezgene) %>% dplyr::select(term, gene) %>% distinct


enrichment_data_by_score = data.frame('Module' = as.character(), 'size' = as.numeric(), 
                                      'SFARI_score' = as.character(), 'pval_ORA' = as.numeric(), 
                                      'padj_ORA' = as.numeric())

SFARI_scores = SFARI_genes$`gene-score` %>% table %>% names

for(i in 1:length(modules)){
  
  module = modules[i]
  genes_in_module = biomart_output$entrezgene[biomart_output$Module==module]
  ORA_module = enricher(gene = genes_in_module, universe = biomart_output$entrezgene %>% as.character, 
                        pAdjustMethod = 'bonferroni', TERM2GENE = term2gene, 
                        pvalueCutoff = 1, qvalueCutoff = 1, maxGSSize = 1e3)
  
  if(length(ORA_module)>0){
    ORA_module = ORA_module %>% data.frame %>% dplyr::select(ID, pvalue, p.adjust) %>%
                 add_column(.before = 'ID', Module = module) %>% 
                 add_column(.before = 'ID', size = length(genes_in_module)) %>%
                 dplyr::rename('SFARI_score' = ID, 'pval_ORA' = pvalue, 'padj_ORA' = p.adjust)
    
    missing_scores = SFARI_scores[!SFARI_scores %in% ORA_module$SFARI_score]
  } else{
    missing_scores = SFARI_scores
  }
  
  if(length(missing_scores)>0){
    for(ms in missing_scores) ORA_module = rbind(ORA_module, c(module, length(genes_in_module), ms, 1, 1))
  }
  
  colnames(ORA_module) = c('Module','size','SFARI_score','pval_ORA','padj_ORA')
  
  enrichment_data_by_score = rbind(enrichment_data_by_score, ORA_module)
}


enrichment_data_by_score = enrichment_data_by_score %>% 
                           left_join(genes_info %>% dplyr::select(Module, MTcor) %>% unique, by = 'Module')


enrichment_data_by_score = enrichment_data_by_score %>% 
                  mutate(color = SFARI_colour_hue(r = 1:3)[SFARI_score %>% as.numeric])

ggplotly(enrichment_data_by_score %>% 
         ggplot(aes(MTcor, 1-as.numeric(pval_ORA), size=size, color = color, fill = color)) + 
         geom_point(alpha=0.5, aes(id=Module)) + geom_smooth(size = 0.5, alpha = 0.1) +
         xlab('Module-Diagnosis Correlation') + ylab('Enrichment') + 
         scale_colour_manual(values = SFARI_colour_hue(r=1:3)) +
         scale_fill_manual(values = SFARI_colour_hue(r=1:3)) + 
         ggtitle('Enrichment by SFARI Gene Score') + 
         theme_minimal() + theme(legend.position = 'none'))

rm(i, module, genes_in_module, ORA_module, term2gene, SFARI_scores, missing_scores, ms, enrichment_data_by_score)
```

<br><br>

## Old SFARI Scores
<br>

```{r warning=FALSE, message=FALSE}
# SFARI Genes
SFARI_genes = read_csv('./../../../SFARI/Data/SFARI_genes_08-29-2019_w_ensembl_IDs.csv')
SFARI_genes = SFARI_genes[!duplicated(SFARI_genes$ID) & !is.na(SFARI_genes$ID),]

plot_data = plot_data %>% dplyr::select(-gene.score) %>% 
            left_join(SFARI_genes %>% dplyr::select(ID, `gene-score`), by = 'ID') %>%
            mutate(gene.score = ifelse(!is.na(`gene-score`), `gene-score`, 
                                       ifelse(ID %in% GO_neuronal$ID, 'Neuronal', 'Others')))

```

### SFARI Scores vs Gene Significance
<br>

- There is no clear pattern here, not a single pairwise comparison is statistically significant


```{r old_SFARI_GS, , fig.height=9, fig.width=10, warning=FALSE, message=FALSE}

comparisons = list(c('1','2'), c('2','3'), c('3','4'), c('4','5'), c('5','6'), c('6','Neuronal'), c('Neuronal','Others'),
                   c('1','3'), c('3','5'), c('5','Neuronal'), 
                   c('2','4'), c('4','6'), c('6','Others'),
                   c('1','4'), c('4','Neuronal'),
                   c('2','5'), c('5','Others'),
                   c('3','Neuronal'), c('1','5'), c('2','6'), c('3','Neuronal'), c('4','Others'))
increase = 0.04
base = 0.75
pos_y_comparisons = c(rep(base, 7), rep(base + increase,3), rep(base + 2*increase,3), rep(base + 3*increase, 2),
                      rep(base + 4*increase, 2), base + 5:9*increase)

plot_data %>% ggplot(aes(gene.score, abs(GS), fill=gene.score)) + 
              geom_boxplot(outlier.colour='#cccccc', outlier.shape='o', outlier.size=3) + 
              stat_compare_means(comparisons = comparisons, label = 'p.signif', method = 't.test', 
                                 method.args = list(var.equal = FALSE), label.y = pos_y_comparisons, 
                                 tip.length = .01) +
              scale_fill_manual(values=SFARI_colour_hue(r=c(1:6,8,7))) + theme_minimal() + 
              ylab('Gene Significance Magnitude') + xlab('SFARI Scores') + theme(legend.position='none')


rm(increase, base, pos_y_comparisons)
```

### SFARI Scores vs Module-Diagnosis correlation
<br>

- Most differences are not statistically significant

```{r old_SFARI_MD, fig.height=9, fig.width=10, warning=FALSE, message=FALSE}

increase = 0.06
base = .95
pos_y_comparisons = c(rep(base, 7), rep(base + increase,3), rep(base + 2*increase,3), rep(base + 3*increase, 2),
                      rep(base + 4*increase, 2), base + 5:9*increase)

plot_data %>% filter(Module != 'gray') %>% ggplot(aes(gene.score, abs(MTcor), fill=gene.score)) + 
              geom_boxplot(outlier.colour='#cccccc', outlier.shape='o', outlier.size=3) + 
              stat_compare_means(comparisons = comparisons, label = 'p.signif', method = 't.test', 
                                 method.args = list(var.equal = FALSE), label.y = pos_y_comparisons, 
                                 tip.length = .01) +
              scale_fill_manual(values=SFARI_colour_hue(r=c(1:6,8,7))) + theme_minimal() + 
              ylab('Module-Diagnosis Correlation Magnitude') + xlab('SFARI Scores') + theme(legend.position='none')


rm(increase, base, pos_y_comparisons)
```


### SFARI Enrichment vs Module-Diagnosis correlation
<br>

#### Percentage of SFARI Genes vs ORA by Module
<br>

```{r, warning=FALSE, message=FALSE}
# Calculate % and ORA of SFARI Genes in each module

# update biomart_output SFARI labels
biomart_output = biomart_output %>% dplyr::select(-`gene-score`) %>%
                 left_join(plot_data %>% dplyr::select(ID, `gene-score`), by = c('ensembl_gene_id'='ID'))

# We need to build a term2gene dataframe with the genes and their SFARI Scores
term2gene = biomart_output %>% mutate(term = ifelse(is.na(`gene-score`), 'Others', 'SFARI'), 
                                      'gene' = entrezgene) %>%  dplyr::select(term, gene) %>% distinct


enrichment_data = data.frame('Module' = modules, 'size' = 0, 'perc_SFARI' = 0,
                       'pval_ORA' = 0, 'padj_ORA' = 0)

for(i in 1:length(modules)){
  module = modules[i]
  genes_in_module = biomart_output$entrezgene[biomart_output$Module==module]
  ORA_module = enricher(gene = genes_in_module, universe = biomart_output$entrezgene %>% as.character, 
                        pAdjustMethod = 'bonferroni', TERM2GENE = term2gene, 
                        pvalueCutoff = 1, qvalueCutoff = 1, maxGSSize = 50000) %>% 
                        data.frame %>% dplyr::select(-geneID,-Description)
  ORA_pval = ifelse('SFARI' %in% ORA_module$ID, ORA_module$pvalue[ORA_module$ID=='SFARI'], 1)
  ORA_padj = ifelse('SFARI' %in% ORA_module$ID, ORA_module$p.adjust[ORA_module$ID=='SFARI'], 1)
  enrichment_data[i,-1] = c(length(genes_in_module), 
                          mean(!is.na(plot_data$`gene-score`[plot_data$Module==module])), 
                          ORA_pval, ORA_padj)
}

enrichment_data = enrichment_data %>% 
                  left_join(genes_info %>% dplyr::select(Module, MTcor) %>% unique, by = 'Module')

rm(i, module, genes_in_module, ORA_module, ORA_pval, ORA_padj, getinfo, mart, term2gene)
```

**Note:** in ORA, a lower probability indicates a higher enrichment in SFARI Genes, so I'm using 1-probability to measure the enrichment of each module
<br>

The two metrics agree quite well, but if we focus on the modules with the highest 'enrichment' with both metrics we can see that the smallest modules are below the trend line while the larger ones are above, which shows the bias towards small modules the percentage of SFARI Genes metric has, making the ORA Enrichment metric more desirable

```{r, warning=FALSE, message=FALSE, fig.align='center'}
ggplotly(enrichment_data %>% ggplot(aes(perc_SFARI, 1-pval_ORA, size=size)) + 
         geom_point(color = enrichment_data$Module, alpha = 0.5, aes(id=Module)) + 
         geom_smooth(color='gray', se = FALSE) + 
         xlab('% of SFARI Genes') + ylab('ORA Enrichment') + ggtitle('Modules') + theme_minimal())
```

#### SFARI Enrichment vs Module-Diagnosis Correlation
<br>

There is a stronger relation between these two metrics than with the new SFARI Genes:

- Mdules with a high negative Module-Diagnosis correlation are the ones with the lowest trend of Enrichment on SFARI Genes<span style="color:red">(!)</span>

- Modules with a Module-Diagnosis correlation close to zero are the ones with the highest Enrichment in SFARI Genes

<br>
Cor(SFARI Enrichment, Module-Diagnosis) = `r round(cor(enrichment_data$MTcor, 1-enrichment_data$pval_ORA),2)[1]`
<br>

- Correlation for modules with a positive Module-Diagnosis correlation = `r round(cor(enrichment_data$MTcor[enrichment_data$MTcor>=0], 1-enrichment_data$pval_ORA[enrichment_data$MTcor>=0]),2)[1]`

- Correlation for modules with a negative Module-Diagnosis correlation = `r round(cor(enrichment_data$MTcor[enrichment_data$MTcor<0], 1-enrichment_data$pval_ORA[enrichment_data$MTcor<0]),2)[1]`

```{r, warning=FALSE, message=FALSE, fig.align='center'}
ggplotly(enrichment_data %>% ggplot(aes(MTcor, 1-pval_ORA, size=size)) + 
         geom_point(color=enrichment_data$Module, alpha=0.5, aes(id=Module)) + 
         geom_smooth(color='#cccccc', size = 0.5, alpha=0.1) + xlab('Module-Diagnosis Correlation') + 
         ggtitle('SFARI Enrichment vs Module-Diagnosis Correlation') +
         ylab('Enrichment in SFARI Genes') + theme_minimal() + theme(legend.position = 'none'))
```

Dividing the SFARI Genes by score we get the same pattern, just much noisier because the groups are smaller now. SFARI Score 1 could be enriched in genes with high (positive) Module-Diagnosis correlation, but this pattern also could be just because there are very few genes with a SFARI Score 1, making the results unreliable


```{r, warning=FALSE, message=FALSE, fig.align='center'}

# We need to build a term2gene dataframe with the genes and their SFARI Scores
term2gene = biomart_output %>%
            mutate('term' = ifelse(`gene-score` == 'Others', 'Others', `gene-score`), 
                   'gene' = entrezgene) %>% dplyr::select(term, gene) %>% distinct


enrichment_data_by_score = data.frame('Module' = as.character(), 'size' = as.numeric(), 
                                      'SFARI_score' = as.character(), 'pval_ORA' = as.numeric(), 
                                      'padj_ORA' = as.numeric())

SFARI_scores = SFARI_genes$`gene-score` %>% table %>% names

for(i in 1:length(modules)){
  
  module = modules[i]
  genes_in_module = biomart_output$entrezgene[biomart_output$Module==module]
  ORA_module = enricher(gene = genes_in_module, universe = biomart_output$entrezgene %>% as.character, 
                        pAdjustMethod = 'bonferroni', TERM2GENE = term2gene, 
                        pvalueCutoff = 1, qvalueCutoff = 1, maxGSSize = 1e3)
  
  if(length(ORA_module)>0){
    ORA_module = ORA_module %>% data.frame %>% dplyr::select(ID, pvalue, p.adjust) %>%
                 add_column(.before = 'ID', Module = module) %>% 
                 add_column(.before = 'ID', size = length(genes_in_module)) %>%
                 dplyr::rename('SFARI_score' = ID, 'pval_ORA' = pvalue, 'padj_ORA' = p.adjust)
    
    missing_scores = SFARI_scores[!SFARI_scores %in% ORA_module$SFARI_score]
  } else{
    missing_scores = SFARI_scores
  }
  
  if(length(missing_scores)>0){
    for(ms in missing_scores) ORA_module = rbind(ORA_module, c(module, length(genes_in_module), ms, 1, 1))
  }
  
  colnames(ORA_module) = c('Module','size','SFARI_score','pval_ORA','padj_ORA')
  
  enrichment_data_by_score = rbind(enrichment_data_by_score, ORA_module)
}


enrichment_data_by_score = enrichment_data_by_score %>% 
                           left_join(genes_info %>% dplyr::select(Module, MTcor) %>% unique, by = 'Module')


enrichment_data_by_score = enrichment_data_by_score %>% 
                  mutate(color = SFARI_colour_hue(r = 1:6)[SFARI_score %>% as.numeric])

ggplotly(enrichment_data_by_score %>% 
         ggplot(aes(MTcor, 1-as.numeric(pval_ORA), size=size, color = color, fill = color)) + 
         geom_point(alpha=0.3, aes(id=Module)) + geom_smooth(size = 0.5, alpha = 0.05) +
         xlab('Module-Diagnosis Correlation') + ylab('Enrichment') + 
         scale_colour_manual(values = SFARI_colour_hue(r=1:6)) +
         scale_fill_manual(values = SFARI_colour_hue(r=1:6)) + 
         ggtitle('Enrichment by SFARI Gene Score') + 
         theme_minimal() + theme(legend.position = 'none'))


rm(i, module, genes_in_module, ORA_module, term2gene, SFARI_scores, missing_scores, ms, enrichment_data_by_score)
```


<br><br>

#### Comparison between Old and New SFARI Genes
<br>

- With both SFARI Gene lists, there seems to be a megative correlation between the magnitude of the Mdoule-Diagnosis correlation and the SFARI Enrichment by module

- The negative enrichment in modules with high negative Module-Diagnosis correlation seems to be a bit stronger with the old SFARI Genes scoring criteria, but they are quite similar

```{r, warning=FALSE, message=FALSE, fig.align='center'}

compare_plot = enrichment_data %>% dplyr::mutate('Old_SFARI' = pval_ORA) %>% 
               dplyr::select(Module, size, MTcor, Old_SFARI) %>%
               left_join(enrichment_data_new_SFARI %>% dplyr::mutate('New_SFARI' = pval_ORA) %>%
                         dplyr::select(Module, New_SFARI), by = 'Module') %>% 
               melt(id.vars = c('Module','size','MTcor')) %>% dplyr::rename('SFARI_version' = variable) %>%
               mutate(size = sqrt(100*as.numeric(size)/max(as.numeric(size))))

p = compare_plot %>% ggplot(aes(MTcor, 1-as.numeric(value), color = SFARI_version, fill = SFARI_version)) +
    geom_point(alpha = 0.5, size=compare_plot$size, aes(id=Module)) + geom_smooth(alpha = 0.1) + 
    xlab('Module-Diagnosis Correlation') + ylab('Enrichment in SFARI Genes') + 
    ggtitle('SFARI Enrichment vs Module-Diagnosis Correlation') +
    theme_minimal() + theme(legend.position = 'bottom')
          
ggExtra::ggMarginal(p, type = 'density', margins = 'y', groupColour = TRUE, groupFill = TRUE, size=10)


rm(p, compare_plot)
```

<br><br>

### Conclusion:
<br>

- SFARI genes not consistent with the other measurements, they even sometimes seem to contradict them, although the strength of the relations is usually not strong enough to be statistically significant

- There is a negative relation between SFARI Enrichment and Module-Diagnosis correlation magnitude, being the modules with negative Module-Diagnosis correlation the most affected

- These patterns are visible in both old and new SFARI Scoring systems, but they are perhaps a bit clearer with the old SFARI scoring system

<br><br>

#### Session info

```{r print_session_info}
sessionInfo()
```
<br><br>
