==> Wright/AllRegions/RMarkdowns/16_DisGeNET_vs_SFARI.Rmd <==
---
title: 'DisGeNET'
output:
  html_document:
    code_folding: 'hide'
---

<br><br>

---

<br>

**NOTE:** I'm just going to include the analysis involving mean expression (since these are the only ones that have different results depending on the gene expression dataset used). The complete analysis can be found in Gandal/AllRegions/RMarkdowns/16_DisGeNET_vs_SFARI.html
<br><br>

---

```{r load_packages, echo=TRUE, warning=FALSE, message=FALSE}
library(tidyverse) ; library(reshape2) ; library(glue) ; library(plotly) ; library(plotlyutils)
library(RColorBrewer) ; library(viridis) ; require(gridExtra) ; library(GGally) ; library(ggpubr)
library(VennDiagram)
library(expss)
library(biomaRt)
library(disgenet2r)
```

```{r load_data, warning=FALSE, message=FALSE}

# Gandal dataset
load('./../Data/preprocessed_data.RData')
datExpr = datExpr %>% data.frame
rownames(datExpr) = datGenes$ensembl_gene_id
DE_info = DE_info %>% data.frame
datMeta = datMeta %>% mutate(ID = title)


# Update gene scores to new SFARI Genes
SFARI_genes = read_csv('./../../../SFARI/Data/SFARI_genes_01-03-2020_w_ensembl_IDs.csv')
SFARI_genes = SFARI_genes[!duplicated(SFARI_genes$ID) & !is.na(SFARI_genes$ID),]


# Add gene symbol
getinfo = c('ensembl_gene_id','external_gene_id')
mart = useMart(biomart='ENSEMBL_MART_ENSEMBL', dataset='hsapiens_gene_ensembl',
               host='feb2014.archive.ensembl.org') ## Gencode v19
gene_names = getBM(attributes=getinfo, filters=c('ensembl_gene_id'), values=rownames(datExpr), mart=mart)

# DisGeNET
DisGeNET = list()
DisGeNET[['asd']] = disease2gene(disease = 'C0004352')@qresult   # Autism Spectrum Disorder
DisGeNET[['scz']] = disease2gene(disease = 'C0036341')@qresult   # Schizophrenia
DisGeNET[['bd']]  = disease2gene(disease = 'C0005586')@qresult   # Bipolar Disorder
DisGeNET[['id']]  = disease2gene(disease = 'C3714756')@qresult   # Intellectual Disability
DisGeNET[['dd']]  = disease2gene(disease = 'C0011581')@qresult   # Depressive Disorder
DisGeNET[['ai']]  = disease2gene(disease = 'C0001973')@qresult   # Alcoholic Intoxication, Chronic
DisGeNET[['bc']]  = disease2gene(disease = 'C0006142')@qresult   # Malignant Carcinoma of Breast
DisGeNET[['cc']]  = disease2gene(disease = 'C0009402')@qresult   # Colorectal Carcinoma
#lc  = disease2gene(disease = 'C0023893')@qresult                # Liver Cirrhosis NO: Only 2 scores
#ht  = disease2gene(disease = 'C0019193')@qresult                # Hepatitis, Toxic No: Only 2 scores
DisGeNET[['dm']]  = disease2gene(disease = 'C0011860')@qresult   # Diabetes Melitus
DisGeNET[['ob']]  = disease2gene(disease = 'C0028754')@qresult   # Obesity
DisGeNET[['ra']]  = disease2gene(disease = 'C0003873')@qresult   # Rheumatoid Arthiritis

SFARI_colour_hue = function(r) {
  pal = c('#FF7631','#FFB100','#E8E328','#b3b3b3','#808080','gray','#d9d9d9')[r]
}


rm(GO_annotations, mart, getinfo, efit, DE_info)
```
<br>

## Mental Disorders
<br>

- Autism Spectrum Disorders (`r nrow(DisGeNET[['asd']])` genes)

- Schizophrenia (`r nrow(DisGeNET[['scz']])`)

- Bipolar Disorder (`r nrow(DisGeNET[['bd']])`)

- Intellectual Disability (`r nrow(DisGeNET[['id']])`)
  
- Depressive Disorder (`r nrow(DisGeNET[['dd']])`)

- Chronic Alcoholic Intoxication (`r nrow(DisGeNET[['ai']])`)

<br><br>

### DisGeNET scores vs Mean Expression
<br>

It's not easy to tell if there's some kind of relation between mean expression and the scores from the different diseases
```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=4}

disgenet_info = data.frame('ID' = gene_names$ensembl_gene_id, 'gene_symbol' = gene_names$external_gene_id, 
                           'meanExpr' = rowMeans(datExpr)) %>% 
                left_join(DisGeNET[['asd']] %>% dplyr::rename('ASD'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['scz']] %>% dplyr::rename('Schizophrenia'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['bd']] %>% dplyr::rename('BipolarDisorder'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['id']] %>% dplyr::rename('IntellectualDisability'=score),by='gene_symbol')%>%
                left_join(DisGeNET[['dd']] %>% dplyr::rename('DepressiveDisorder'=score), by='gene_symbol') %>%
                left_join(DisGeNET[['ai']] %>% dplyr::rename('Alcoholism'=score), by = 'gene_symbol') %>%
                dplyr::select(gene_symbol, meanExpr, ASD, Schizophrenia, BipolarDisorder, 
                              IntellectualDisability, DepressiveDisorder, Alcoholism)


disgenet_info %>% dplyr::select(-gene_symbol) %>% melt(id.vars = 'meanExpr') %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(meanExpr, value, color = Disorder)) + geom_point(alpha=0.3) + 
                  facet_grid(~Disorder) + geom_smooth(alpha = 0.2, method = 'loess') + 
                  xlab('Mean Expression') + ylab('Score') + scale_y_log10() + theme_minimal() + 
                  theme(legend.position = 'none')


rm(asd_info)
```

Separating the genes by SFARI and non-SFARI doesn't help to make the relation clearer, but it does suggest that SFARI Genes have in general higher scores, not only in ASD scores, but also in other neurodevelopmental disorders

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=6}

disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
                  dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(meanExpr, value, color = SFARI)) + geom_point(alpha=0.3) + 
                  facet_grid(~Disorder) + geom_smooth(alpha = 0.2, method = 'loess') + 
                  xlab('Mean Expression') + ylab('Score') + scale_y_log10() + theme_minimal() + 
                  theme(legend.position = 'bottom')

```
<br>

### DisGeNET binned scores vs Mean Expression
<br>

As we saw above, the relation is sometimes not that clear in the continuous scale (maybe because there aren't that many points and the relation is not that strong?) I'm going to separate the scores into three groups to see if a pattern is more visible this way

Binned scores (selected to balance the number of genes in each group): 

- Low: (0,0.31]

- Medium: (0.31,0.45]

- High: (0.45,1]

<br>

There doesn't seem to be a relation between level of expression and the binned scores for any of the diseases. This is weird because in Gandal's dataset we found there was a statistically significant difference
<br>

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=4}

comparisons = list(c('Low','Medium'), c('Medium','High'), c('Low','High'))
increase = 1.2
base = 11
pos_y_comparisons = c(base, base, base + increase)

disgenet_info %>% dplyr::select(-gene_symbol) %>% melt(id.vars = 'meanExpr') %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                             labels = c('Low','Medium','High'))) %>%
                  mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low'))) %>%
                  ggplot(aes(binned_scores, meanExpr)) + facet_grid(~Disorder) +
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = Disorder)) + 
                  stat_compare_means(comparisons = comparisons, label = 'p.signif', method = 't.test', 
                                     method.args = list(var.equal = FALSE), label.y = pos_y_comparisons, 
                                     tip.length = 0.015) +
                  xlab('Binned Scores') + ylab('Mean Expression') + 
                  theme_minimal() + theme(legend.position = 'none')


rm(asd_info, comparisons, increase, base, pos_y_comparisons)
```
<br>

When separating the genes that belong to the SFARI Genes, the bias in mean expression in the genes related to Depressive Disorder and in SFARI Genes becomes statistically significant
```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=6}

comparisons = list(c('Low','Medium'), c('Medium','High'), c('Low','High'))
increase = 1.5
base = 11
pos_y_comparisons = c(base, base, base + increase)

disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
                  dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                             labels = c('Low','Medium','High'))) %>%
                  mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low'))) %>%
                  ggplot(aes(binned_scores, meanExpr)) + facet_grid(SFARI ~ Disorder) +
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = Disorder)) + 
                  stat_compare_means(comparisons = comparisons, label = 'p.signif', method = 't.test', 
                                  method.args = list(var.equal = FALSE), label.y = pos_y_comparisons, 
                                  tip.length = 0.015) +
                  xlab('Binned Scores') + ylab('Mean Expression') +
                  theme_minimal() + theme(legend.position = 'none')


rm(comparisons, increase, base, pos_y_comparisons)
```
<br><br>

---

<br><br>

## Other Diseases
<br>

- Malignant neoplasm of breast (1074)

- Experimental Liver Cirrhosis (774)

- Colorectal carcinoma (702)

- Hepatitis, Toxic (404)

- Diabetes Melitus (221)

<br>

### DisGeNET scores vs Mean Expression
<br>

As before, this continuous scale for DisGeNET scores doesn't allow us to see clearly if there is a relation or not between these two variables

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=4}

disgenet_info = data.frame('ID' = gene_names$ensembl_gene_id, 'gene_symbol' = gene_names$external_gene_id, 
                       'meanExpr' = rowMeans(datExpr)) %>% 
                left_join(DisGeNET[['bc']] %>% dplyr::rename('BreastCancer'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['cc']] %>% dplyr::rename('ColorectalCarcinoma'=score), by='gene_symbol')%>%
                left_join(DisGeNET[['dm']] %>% dplyr::rename('Diabetes'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['ob']] %>% dplyr::rename('Obesity'=score), by = 'gene_symbol') %>%
                left_join(DisGeNET[['ra']] %>% dplyr::rename('Arthritis'=score), by = 'gene_symbol') %>%
                dplyr::select(gene_symbol, meanExpr, BreastCancer, ColorectalCarcinoma, Diabetes, Obesity,
                              Arthritis)


disgenet_info %>% dplyr::select(-gene_symbol) %>% melt(id.vars = 'meanExpr') %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  ggplot(aes(meanExpr, value, color = Disorder)) + geom_point(alpha=0.3) + 
                  facet_grid(~Disorder) + geom_smooth(alpha = 0.2, method = 'loess') + 
                  xlab('Mean Expression') + ylab('Score') + scale_y_log10() + theme_minimal() + 
                  theme(legend.position = 'none')

```
<br>

### DisGeNET binned scores vs Mean Expression
<br>

Usin the same binned scores as before: 

- Low: (0,0.31]

- Medium: (0.31,0.45]

- High: (0.45,1]

<br>

Perhaps there could be a relation between the binned scores and mean expression in breast cancer, but the p-value is not that small
<br>

```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=4}

comparisons = list(c('Low','Medium'), c('Medium','High'), c('Low','High'))
increase = 1.2
base = 11
pos_y_comparisons = c(base, base, base + increase)

disgenet_info %>% dplyr::select(-gene_symbol) %>% melt(id.vars = 'meanExpr') %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                             labels = c('Low','Medium','High'))) %>%
                  mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low'))) %>%
                  ggplot(aes(binned_scores, meanExpr)) + facet_grid(~Disorder) +
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill = Disorder)) + 
                  stat_compare_means(comparisons = comparisons, label = 'p.signif', method = 't.test', 
                                     method.args = list(var.equal = FALSE), label.y = pos_y_comparisons, 
                                     tip.length = 0.015) +
                  xlab('Binned Scores') + ylab('Mean Expression') + 
                  theme_minimal() + theme(legend.position = 'none')


rm(asd_info, comparisons, increase, base, pos_y_comparisons)
```
<br>

Separating SFARI Genes
```{r, warning=FALSE, message=FALSE, fig.width=10, fig.height=6}

comparisons = list(c('Low','Medium'), c('Medium','High'), c('Low','High'))
increase = 1.5
base = 11
pos_y_comparisons = c(base, base, base + increase)

disgenet_info %>% mutate('SFARI' = gene_symbol %in% SFARI_genes$`gene-symbol`) %>% 
                  dplyr::select(-gene_symbol) %>% melt(id.vars = c('meanExpr', 'SFARI')) %>% 
                  dplyr::rename('Disorder' = variable) %>% filter(!is.na(value)) %>%
                  mutate(binned_scores = cut(value, breaks = c(0, 0.31, 0.45, 1), 
                                             labels = c('Low','Medium','High'))) %>%
                  mutate(binned_scores = factor(binned_scores, levels = c('High','Medium','Low'))) %>%
                  ggplot(aes(binned_scores, meanExpr)) + facet_grid(SFARI ~ Disorder) +
                  geom_boxplot(outlier.colour='gray', outlier.shape='o', outlier.size=3, aes(fill=Disorder)) + 
                  stat_compare_means(comparisons = comparisons, label = 'p.signif', method = 't.test', 
                                  method.args = list(var.equal = FALSE), label.y = pos_y_comparisons, 
                                  tip.length = 0.015) +
                  xlab('Binned Scores') + ylab('Mean Expression') + 
                  theme_minimal() + theme(legend.position = 'none')

rm(comparisons, increase, base, pos_y_comparisons)
```
<br><br>

---

<br><br>

## Conclusions
<br>

There is a significant relation between level of expression and DisGeNET scores in ASD and Depressive Disorder, but the last one, only when selecting the SFARI Genes

We didn't find a strong relation between level of expression in any of the diseases as we had found in Gandal's dataset

<br><br>
